\name{Cul}
\alias{Cul}
\title{Cochran C critical value}
\usage{
Cul(a,n,N)
}
\arguments{
  \item{a}{significance level}

  \item{n}{number of points per series}

  \item{N}{number of data series}
}
\description{
Upper limit critical value Cul for one-sided test on
balanced design
}
\author{
Antoine Stevens
}
\references{
\url{http://www.en.wikipedia.org/wiki/Cochran's_C_test}
}
\keyword{internal}

