\name{fastDistV}
\alias{fastDistV}
\title{A fast distance algorithm for a matrix and a vector written in C++}
\usage{
  fastDistV(X,Y,method)

  fastDistV(X,Y,method)
}
\arguments{
  \item{X}{a \code{matrix}}

  \item{Y}{a \code{vector}}

  \item{method}{a \code{string} with possible values
  "euclid", "cor", "cosine"}

  \item{X}{a \code{matrix}}

  \item{Y}{a \code{vector}}

  \item{method}{a \code{string} with possible values
  "euclid", "cor", "cosine"}
}
\value{
  a distance \code{vector}

  a distance \code{vector}
}
\description{
  A fast distance algorithm for a matrix and a vector
  written in C++

  A fast distance algorithm for a matrix and a vector
  written in C++
}
\author{
  Antoine Stevens and Leonardo Ramirez-Lopez

  Antoine Stevens and Leonardo Ramirez-Lopez
}
\keyword{internal}

