% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nircal.R
\name{get_nircal_comments}
\alias{get_nircal_comments}
\title{get the comments of the spectra in the nircal file}
\usage{
get_nircal_comments(connection, metanumbers, begin_s, comment_s, comment_f, n)
}
\description{
internal
}
\keyword{internal}
