% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nircal.R
\name{get_nircal_description}
\alias{get_nircal_description}
\title{get the description of the spectra in the nircal file}
\usage{
get_nircal_description(x, begin_s, spcinfo, comment_s, comment_f, n)
}
\description{
internal
}
\keyword{internal}
