% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nircal.R
\name{get_nircal_lengthspc}
\alias{get_nircal_lengthspc}
\title{get the number of spectral variables in the nircaa file}
\usage{
get_nircal_lengthspc(connection, from, to)
}
\description{
internal
}
\keyword{internal}
