% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nircal.R
\name{get_nircal_metadata}
\alias{get_nircal_metadata}
\title{get the metadata of the samples in the nircal file}
\usage{
get_nircal_metadata(
  connection,
  n,
  spctra_start,
  spcinfo,
  progress,
  pb,
  progress.start,
  progress.steps
)
}
\description{
internal
}
\keyword{internal}
