% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nircal.R
\name{get_nircal_spectra}
\alias{get_nircal_spectra}
\title{get the spectra in the nircal file}
\usage{
get_nircal_spectra(x, values_s, spctra_start, speclength, n)
}
\description{
internal
}
\keyword{internal}
