% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_info.R
\name{pkg_info}
\alias{pkg_info}
\title{Get the package version info}
\usage{
pkg_info(pkg = "prospectr")
}
\arguments{
\item{pkg}{the package name i.e "prospectr"}
}
\description{
returns package info.
}
\keyword{internal}
