% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedsnv.R
\name{wsnv}
\alias{wsnv}
\title{Weighted SNV or Variable Sorting for Normalization (VSN) with SNV}
\usage{
wsnv(x, tol = 0.001, nbParam = 3, nbIter = 5000, progress = FALSE)
}
\arguments{
\item{x}{a \code{matrix} - the spectral matrix (each spectrum is a row).}

\item{tol}{a \code{numeric(1)} - error tolerance threshold for solving the equations (default value: 0.001).}

\item{nbParam}{an \code{integer(1)} - number of parameter considered in the equations (can be 2, 3 or 4).}

\item{nbIter}{an \code{numeric(1)} - number of iteration for the RANSAC algorithm (default value: 5000).}

\item{progress}{a \code{logical} - if TRUE, algorithm progress is shown in the console.}
}
\value{
a \code{matrix} of corrected spectra.
}
\description{
Function used to normalize the spectra and that computes the optimal wavelength weighting by considering
all possible spectrum couples (RANSAC algorithm) and summing their results.
}
\examples{
NIRsoil_wsnv<-wsnv(NIRsoil$spc[1:5,])

}
\references{
Rabatel, G., Marini, F., Walczak, B. & Roger, J.-M. (2020) VSN: Variable sorting for normalization.
Journal of Chemometrics 34:e3164. DOI: \url{https://doi.org/10.1002/cem.3164}
}
\seealso{
\code{\link{snv}} \code{\link{wmsc}}
}
\author{
\itemize{
\item Original MATLAB code: Jean-Michel Roger.
\item Conversion to R & Parallelization: Guillaume Hans (with permission of the author).
}
}
