% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cul.R
\name{Cul}
\alias{Cul}
\title{Cochran C critical value}
\usage{
Cul(a,n,N)
}
\arguments{
\item{a}{significance level.}

\item{n}{number of points per series.}

\item{N}{number of data series.}
}
\description{
Upper limit critical value Cul for one-sided test on balanced design
}
\references{
\url{https://en.wikipedia.org/wiki/Cochran's_C_test}
}
\author{
Antoine Stevens
}
\keyword{internal}
