\name{fastDist}
\alias{fastDist}
\title{A fast distance algorithm for two matrices written in C++}
\usage{
fastDist(X,Y,method)
}
\arguments{
  \item{X}{a \code{matrix}}

  \item{Y}{a \code{matrix}}

  \item{method}{a \code{string} with possible values
  "euclid", "cor", "cosine"}
}
\value{
a distance \code{matrix}
}
\description{
A fast distance algorithm for two matrices written in C++
}
\author{
Antoine Stevens and Leonardo Ramirez-Lopez
}
\keyword{internal}

