% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_msc_coeff}
\alias{get_msc_coeff}
\title{get_msc_coeff}
\usage{
get_msc_coeff(X, ref_spectrum)
}
\arguments{
\item{X}{matrix}

\item{ref_spectrum}{a matrix of one row and same columns as in X}
}
\description{
Coefficients for multiplicative Scatter Correction written in C++
}
\keyword{internal}
