% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc.R
\name{msc}
\alias{msc}
\title{MSC (Multiplicative Scatter Correction)}
\usage{
msc(x)
}
\arguments{
\item{x}{a \code{matrix} - the spectral matrix (each spectrum is a row).}
}
\value{
a \code{matrix} of corrected spectra.
}
\description{
The \code{msc} function implements the multiplicative scatter correction based on the mean spectrum.
}
\examples{
bioSpec_msc<-msc(bioSpec)

}
\references{
Geladi, P., MacDougall, D., and Martens, H. (1985) Linearization and Scatter-Correction for Near-Infrared Reflectance Spectra of Meat.
Applied Spectroscopy, 39(3):491-500.
}
\seealso{
\code{\link{snv}} \code{\link{mean}}
}
\author{
Guillaume Hans
}
