% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospectr.R
\docType{package}
\name{prospectr-package}
\alias{prospectr-package}
\alias{resemble}
\title{Overview of the functions in the prospectr package}
\description{
This package implements a number of \code{R} functions useful for 
pre-processing spectral spectra (e.g. NIR, IR) as well as for 
selecting representaive samples/spectra .
}
\details{
Currently, the following preprocessing functions are available:

\itemize{ 
  \item{\code{\link{continuumRemoval}}} 
  \item{\code{\link{savitzkyGolay}}} 
  \item{\code{\link{detrend}}} 
  \item{\code{\link{gapDer}}} 
  \item{\code{\link{movav}}} 
  \item{\code{\link{standardNormalVariate}}}
  \item{\code{\link{binning}}}   
  \item{\code{\link{resample}}} 
  \item{\code{\link{resample2}}} 
  \item{\code{\link{blockScale}}}  
  \item{\code{\link{blockNorm}}}
 }

The selection of samples/observations for calibration of \acronym{vis-NIR} data can be achieved with one of the following functions:

\itemize{ 
  \item{\code{\link{naes}}} 
  \item{\code{\link{honigs}}} 
  \item{\code{\link{shenkWest}}}   
  \item{\code{\link{kenStone}}} 
  \item{\code{\link{duplex}}} 
  \item{\code{\link{puchwein}}}  
 } 
 
 Other useful functions are also available:
 
 \itemize{
  \item{\code{\link{read_nircal}}} 
  \item{\code{\link{readASD}}} 
  \item{\code{\link{spliceCorrection}}}
  \item{\code{\link{cochranTest}}}
 }
}
\author{
Antoine Stevens 
Leonardo Ramirez-Lopez \email{ramirez.lopez.leo@gmail.com},
}
