% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospectr.R
\docType{package}
\name{prospectr-package}
\alias{prospectr-package}
\alias{prospectr}
\title{Overview of the functions in the prospectr package}
\description{
\lifecycle{maturing}

Misc functions for spectral data
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package implements a number of \code{R} functions useful for
pre-processing spectral data well as for selecting representaive samples/spectra.
The functions included here are particularly useful in Near-Infrarred and Infrared
Spectroscopy applications.
}
\details{
Currently, the following preprocessing functions are available:

\itemize{
\item{\code{\link{resample}}}
\item{\code{\link{resample2}}}
\item{\code{\link{movav}}}
\item{\code{\link{standardNormalVariate}}}
\item{\code{\link{msc}}}
\item{\code{\link{detrend}}}
\item{\code{\link{blockScale}}}
\item{\code{\link{blockNorm}}}
\item{\code{\link{binning}}}
\item{\code{\link{savitzkyGolay}}}
\item{\code{\link{gapDer}}}
\item{\code{\link{continuumRemoval}}}
}

For the selection of representative samples/observations for calibrating
spectral models the following functions ca be used:

\itemize{
\item{\code{\link{naes}}}
\item{\code{\link{honigs}}}
\item{\code{\link{shenkWest}}}
\item{\code{\link{kenStone}}}
\item{\code{\link{duplex}}}
\item{\code{\link{puchwein}}}
}

Other useful functions are also available:

\itemize{
\item{\code{\link{read_nircal}}}
\item{\code{\link{readASD}}}
\item{\code{\link{spliceCorrection}}}
\item{\code{\link{cochranTest}}}
}
}
\author{
Antoine Stevens & \href{https://orcid.org/0000-0002-5369-5120}{Leonardo Ramirez-Lopez} \email{ramirez.lopez.leo@gmail.com},
}
