% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceCorrection.R
\name{spliceCorrection}
\alias{spliceCorrection}
\title{Splice correction of a spectral matrix}
\usage{
spliceCorrection(X, wav, splice = c(1000, 1830), interpol.bands = 10)
}
\arguments{
\item{X}{A numeric matrix of spectra or a vector representing a single
spectrum to be corrected (optionally a data frame that can be coerced to a
numeric matrix). Rows represent each spectrum and columns the band positions.}

\item{wav}{A numeric vector representing the band positions (e.g. wavelengths)
corresponding to the spectral data.}

\item{splice}{A numeric vector of length 1 or 2 specifying the band
(e.g. wavelength) positions where splicing occurs.
The default values \code{c(1000, 1830)} refer to typical splice positions for
NIR spectrometers like the ASD FieldSpec Pro.
For other spectrometers, different splice positions may be applicable (see
Details).}

\item{interpol.bands}{The number of bands around the splice positions to use
for linear interpolation. Default is 10 bands.}
}
\value{
A matrix containing the spectra with corrected splices, ensuring
smoother transitions across detectors.
}
\description{
Corrects discontinuities (splices) in a spectral matrix by performing linear
interpolation at the boundaries between different detectors in the spectrometer.
This method is commonly used for near-infrared (NIR) spectra where splicing
artifacts occur due to the transition between multiple detectors in the
spectrometer.
}
\details{
This function addresses splicing artifacts caused by transitioning between
different detectors, often seen in NIR spectrometers.
Splicing occurs when data from multiple detectors are combined, potentially
introducing steps or discontinuities in the spectral data.
This correction method uses linear interpolation across a specified number
of bands around the splice points to smooth transitions and eliminate these
artifacts.

The default splice positions are set for the ASD FieldSpec Pro spectrometer,
which has detectors at 1000 nm and 1830 nm.
For other NIR instruments, common splice positions include:
\itemize{
\item{XDS (FOSS): 1100 nm}
\item{ProxiMate (BUCHI Labortechnik): 900 nm}
}
}
\author{
Antoine Stevens and
\href{https://orcid.org/0000-0002-5369-5120}{Leonardo Ramirez-Lopez}
}
